LPS-150989 Extract common code to util class
